Add greeting, add transition to Quote
Add greeting, depends on a time of a day it will display
'Good morning.', 'Good afternoon.', 'Good evening.'.
It will shuffle greeting with a mantra and display random mantra.

Quotes now have transition effect on hover.
Author is hidden, it is visible again on hover effect.
Added hyperlink to wikipedia when clicked on author.